ELEC3848_ProposedFunction-main/requirements.txt
python-dotenv
pyserial
openai==2.7.1
SpeechRecognition==3.14.3
gTTS==2.5.4
langchain-core
requests
